% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cleanup.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{The dataframe constructed by function 'eq_get_data'}
}
\value{
This function returns the cleanedup NOAA earthquake data in a dataframe
}
\description{
This function cleans the NOAA earthquake data extracted from https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1
which was placed in a dataframe by function \code{eq_get_data}. The cleanup consists of:
}
\details{
1. Making LONGITUDE and LATITUDE numeric
2. Make a new dataitem DATE using the dataitems YEAR, MONTH and DAY from the dataframe
3. Cleanup the locationname via function \code{eq_clean_data} by deleting all text following a ':'
}
\examples{
\dontrun{
data <- readr::read_delim("inst/extdata/signif.txt.tsv", delim = "\\t")
eq_clean_data(data)
}

}
