% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A dataframe containing cleanedup NOAA earthquake data}
}
\value{
A character containing HTML text with (if not empty) location, magnitude and deaths of an earthquake
}
\description{
This function creates a HTML tekst which is used as label for a leaflet. The label consists of location,
magnitude and deaths from the cleanedup NOAA earthquake data
}
\examples{
\dontrun{
eq_create_label(data)
}

}
