% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cleanup.R
\name{eq_get_data}
\alias{eq_get_data}
\title{eq_get_data}
\usage{
eq_get_data(datafile)
}
\arguments{
\item{datafile}{A character string denoting the (compressed) file name}
}
\value{
This function returns the input data in a dataframe table
}
\description{
This function reads a datafile delimited by tabs
}
\examples{
\dontrun{

eq_get_data("data/signif.txt.tsv")
}

}
