% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cleanup.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(locationname)
}
\arguments{
\item{locationname}{The column LOCATION_NAME from the NOAA earthquake data}
}
\value{
This function returns the cleanedup LOCATION_NAME
}
\description{
This function cleans the location form the NOAA earthquake data where the earthquake took place.
The cleanup of the location is done by ruling out all text folowed by a ':' data extracted
from https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1
}
\examples{
\dontrun{
eq_clean_data(eq_get_data("data/signif.txt.tsv"))
}

}
