% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A dataframe containing cleanedup NOAA earthquake data}

\item{annot_col}{A character containing the name of the column in the data where the value within the column is used as text in the popup}
}
\value{
A leaflet map with earthquakes and corresponding text.
}
\description{
This function created a leaflet map with location of earthquakes with a popup with selected information of the earthquake
}
\examples{
\dontrun{
eq_map(data, annot_col = "DATE")
}
}
