% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geoms.R
\docType{data}
\name{GeomMyTimeline}
\alias{GeomMyTimeline}
\title{Geom of a timeline containing earthquakes}
\format{An object of class \code{GeomMyTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomMyTimeline
}
\description{
This geom plots a timeline showing earthquakes with the time it occured, the scaled size, the country and the number of casualties momentfunction uses the data file with fatal injuries in motor vehicle traffic crashes in the USA to summarizes the total number
}
\examples{
\dontrun{
data \%>\% eq_clean_data() \%>\%
   filter(COUNTRY \%in\% c("USA"), YEAR > 2000) \%>\%
   ggplot(aes(x = DATE, y = COUNTRY, colour = TOTAL_DEATHS, size = EQ_PRIMARY)) +
   geom_timeline() +
   theme_timeline() +
   labs(size = "Richter scale value", colour = "# deaths")
}

}
\keyword{datasets}
