% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geoms.R
\name{theme_timeline}
\alias{theme_timeline}
\title{theme_timeline}
\usage{
theme_timeline()
}
\description{
This is a function that makes some changes to visualize the eartquakes on the timeline
}
\examples{
\dontrun{
data \%>\% eq_clean_data() \%>\%
filter(COUNTRY \%in\% c("USA"), YEAR > 2000) \%>\%
   ggplot(aes(x = DATE, y = COUNTRY, colour = TOTAL_DEATHS, size = EQ_PRIMARY)) +
   geom_timeline() +
   theme_timeline()
}

}
